requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==22.12.0
click==8.1.3
Django==4.1.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
flake8==6.0.0
inflection==0.5.1
jsonschema==4.17.3
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.2
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
pyrsistent==0.19.3
pytz==2022.7
PyYAML==6.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
uritemplate==4.1.1
